numeric extension move into core part 5
    moved numeric/kernel.hpp into core
