+++ example/convolve2d.cpp
@@ -9,7 +9,7 @@
 
 #include <vector>
 #include <iostream>
-#include <boost/gil/extension/numeric/kernel.hpp>
+#include <boost/gil/image_processing/kernel.hpp>
 #include <boost/gil/image_processing/convolve.hpp>
 #include <boost/gil/extension/io/png.hpp>
