+++ include/boost/gil/image_processing/filter.hpp
@@ -10,7 +10,7 @@
 #ifndef BOOST_GIL_IMAGE_PROCESSING_FILTER_HPP
 #define BOOST_GIL_IMAGE_PROCESSING_FILTER_HPP
 
-#include <boost/gil/extension/numeric/kernel.hpp>
+#include <boost/gil/image_processing/kernel.hpp>
 
 #include <boost/gil/image_processing/convolve.hpp>
