+++ include/boost/gil/image_processing/morphology.hpp
@@ -8,7 +8,8 @@
 
 #ifndef BOOST_GIL_IMAGE_PROCESSING_MORPHOLOGY_HPP
 #define BOOST_GIL_IMAGE_PROCESSING_MORPHOLOGY_HPP
-#include <boost/gil/extension/numeric/kernel.hpp>
+
+#include <boost/gil/image_processing/kernel.hpp>
 #include <boost/gil/gray.hpp>
 #include <boost/gil/image_processing/threshold.hpp>
