+++ test/core/image_processing/CMakeLists.txt
@@ -26,7 +26,9 @@ foreach(_name
     convolve
     convolve_2d
     convolve_cols
-    convolve_rows)
+    convolve_rows
+    kernel
+    kernel_fixed)
     set(_test t_core_image_processing_${_name})
     set(_target test_core_image_processing_${_name})
