+++ test/core/image_processing/Jamfile
@@ -10,6 +10,7 @@
 import testing ;
 
 compile-fail threshold_color_spaces_not_compatible_fail.cpp ;
+compile-fail kernel_1d_fixed_even_size_fail.cpp ;
 run threshold_binary.cpp ;
 run threshold_truncate.cpp ;
 run threshold_otsu.cpp ;
@@ -28,3 +29,5 @@ run convolve.cpp ;
 run convolve_2d.cpp ;
 run convolve_cols.cpp ;
 run convolve_rows.cpp ;
+run kernel.cpp ;
+run kernel_fixed.cpp ;
