+++ test/extension/numeric/Jamfile
@@ -12,10 +12,8 @@ import testing ;
 
 alias headers : [ generate_self_contained_headers extension/numeric ] ;
 
-compile-fail kernel_1d_fixed_even_size_fail.cpp ;
 
-run kernel.cpp ;
-run kernel_fixed.cpp ;
+
 run matrix3x2.cpp ;
 run resample.cpp ;
